requirements.txt
PyQt6
json5
Pillow